requirements.txt
flask
mysql-connector
python-dotenv